backend/requirements.txt
numpy
fastapi
uvicorn
requests
sqlalchemy
schedule
pandas
prophet
scikit-learn
prometheus-client
kubernetes
boto3==1.34.0
botocore==1.34.0
python-dotenv==1.0.0
azure-identity==1.15.0
azure-mgmt-compute==30.3.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-monitor==6.0.1
azure-mgmt-costmanagement==3.0.0
azure-mgmt-resource==23.0.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cryptography
email-validator